move video components to its directory
