+++ src/blocks/video-maxi/components/index.js
@@ -1,0 +1,5 @@
+export { default as PopupSettingsControl } from './popup-settings-control';
+export { default as VideoControl } from './video-control';
+export { default as VideoIconControl } from './video-icon-control';
+export { default as VideoOptionsControl } from './video-options-control';
+export { default as VideoOverlayControl } from './video-overlay-control';
