+++ src/components/index.js
@@ -57,7 +57,6 @@ export { default as OpacityControl } from './opacity-control';
 export { default as OverflowControl } from './overflow-control';
 export { default as PaddingControl } from './padding-control';
 export { default as Placeholder } from './placeholder';
-export { default as PopupSettingsControl } from './popup-settings-control';
 export { default as PositionControl } from './position-control';
 export { default as RawHTML } from './raw-html';
 export { default as RelationControl } from './relation-control';
@@ -82,9 +81,5 @@ export { default as Toolbar } from './toolbar';
 export { default as TransformControl } from './transform-control';
 export { default as TransitionControl } from './transition-control';
 export { default as TypographyControl } from './typography-control';
-export { default as VideoControl } from './video-control';
-export { default as VideoIconControl } from './video-icon-control';
-export { default as VideoOptionsControl } from './video-options-control';
-export { default as VideoOverlayControl } from './video-overlay-control';
 export { default as ZIndexControl } from './zindex-control';
 export * from './svg-color';
